+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # IntelliJ-git-open Changelog
 
-## [Unreleased]
+## 2024.6.3
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Support ssh protocol for git remote url
+- Support Open Pipelines/Actions On Browser